Move httpValidator and tests to resources files
