+++ internal/resources/connections/resources.go
@@ -1,12 +1,16 @@
 package connections
 
 import (
+	"context"
 	"fmt"
+	"net/url"
+
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 
 	"github.com/grafana/terraform-provider-grafana/v3/internal/common"
 	"github.com/grafana/terraform-provider-grafana/v3/internal/common/connectionsapi"
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
 )
 
 var DataSources = []*common.DataSource{
@@ -15,6 +19,52 @@ var DataSources = []*common.DataSource{
 
 var Resources = []*common.Resource{
 	makeResourceMetricsEndpointScrapeJob(),
+}
+
+type HTTPSURLValidator struct{}
+
+func (v HTTPSURLValidator) Description(ctx context.Context) string {
+	return v.MarkdownDescription(ctx)
+}
+
+func (v HTTPSURLValidator) MarkdownDescription(_ context.Context) string {
+	return "value must be valid URL with HTTPS"
+}
+
+func (v HTTPSURLValidator) ValidateString(ctx context.Context, request validator.StringRequest, response *validator.StringResponse) {
+	value := request.ConfigValue.ValueString()
+
+	if value == "" {
+		response.Diagnostics.AddAttributeError(
+			request.Path,
+			v.Description(ctx),
+			"A valid URL is required.\n\n"+
+				fmt.Sprintf("Given Value: %q\n", value),
+		)
+		return
+	}
+
+	u, err := url.Parse(value)
+	if err != nil {
+		response.Diagnostics.AddAttributeError(
+			request.Path,
+			v.Description(ctx),
+			"A string value was provided that is not a valid URL.\n\n"+
+				"Given Value: "+value+"\n"+
+				"Error: "+err.Error(),
+		)
+		return
+	}
+
+	if u.Scheme != "https" {
+		response.Diagnostics.AddAttributeError(
+			request.Path,
+			v.Description(ctx),
+			"A URL was provided, protocol must be HTTPS.\n\n"+
+				fmt.Sprintf("Given Value: %q\n", value),
+		)
+		return
+	}
 }
 
 func withClientForResource(req resource.ConfigureRequest, resp *resource.ConfigureResponse) (*connectionsapi.Client, error) {
